requirements.txt
asgiref==3.8.1
contourpy==1.3.0
cycler==0.12.1
Django==5.1.2
django-cleanup==9.0.0
django-debug-toolbar==4.4.6
django-extensions==3.2.3
django-redis==5.4.0
djangorestframework==3.15.2
factory_boy==3.3.1
Faker==30.8.2
fonttools==4.54.1
kiwisolver==1.4.7
matplotlib==3.9.2
mysqlclient==2.2.5
numpy==2.1.3
packaging==24.1
pillow==11.0.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
redis==5.2.0
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
